+++ src/main/java/org/ivan/artshow/module/artcategory/pojo/dto/ArtcategoryDTO.java
@@ -9,11 +9,11 @@ package org.ivan.artshow.module.artcategory.pojo.dto;
  * @since 1.0.0
  */
 public record ArtcategoryDTO(
-        Integer categoryId,
+        Long categoryId,
         String categoryName,
         String iconUrl
 ) {
-    public Integer getCategoryId() { return categoryId; }
+    public Long getCategoryId() { return categoryId; }
     public String getCategoryName() { return categoryName; }
     public String getIconUrl() { return iconUrl; }
 }
